requirements.txt
# aiohappyeyeballs==2.4.6
# aiohttp==3.10.11
# aiohttp-retry==2.9.1
# aiosignal==1.3.2
# aiosqlite==0.21.0
# alembic==1.14.1
# altair==4.2.2
# amqp==5.3.1
# annotated-types==0.7.0
# antlr4-python3-runtime==4.9.3
# anyio==4.8.0
# # apache-airflow==2.7.2
# # apache-airflow-providers-common-compat==1.5.0
# # apache-airflow-providers-common-io==1.5.0
# # apache-airflow-providers-common-sql==1.23.0
# # apache-airflow-providers-fab==1.5.3
# # apache-airflow-providers-ftp==3.12.2
# # apache-airflow-providers-http==5.2.0
# # apache-airflow-providers-imap==3.8.2
# # apache-airflow-providers-smtp==2.0.0
# # apache-airflow-providers-sqlite==4.0.0
# apispec==6.8.1
# appdirs==1.4.4
# argcomplete==3.5.3
# asgiref==3.8.1
# asyncssh==2.20.0
# atpublic==5.1
# attrs==25.1.0
# babel==2.17.0
# backoff==2.2.1
# beautifulsoup4==4.13.3
# billiard==4.2.1
# blinker==1.9.0
# cachelib==0.13.0
# cachetools==5.5.2
# celery==5.4.0
# certifi==2025.1.31
# cffi==1.17.1
# charset-normalizer==3.4.1
# click==8.1.8
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# clickclick==20.10.2
# colorama==0.4.6
# colorlog==4.7.2
# configobj==5.0.9
# ConfigUpdater==3.2
# connexion==2.14.2
# cron-descriptor==1.4.5
# croniter==6.0.0
# cryptography==44.0.2
# decorator==5.2.1
# Deprecated==1.2.18
# dictdiffer==0.9.0
# dill==0.3.9
# diskcache==5.6.3
# distro==1.9.0
# dnspython==2.7.0
# dotenv==0.9.9
# dpath==2.2.0
# dulwich==0.22.7
# dvc==3.59.1
# dvc-data==3.16.9
# dvc-gs==3.0.1
# dvc-http==2.32.0
# dvc-objects==5.1.0
# dvc-render==1.0.2
# dvc-studio-client==0.21.0
# dvc-task==0.40.2
# email_validator==2.2.0
# entrypoints==0.4
# fairlearn==0.12.0
# filelock==3.17.0
# # Flask==2.2.5
# # Flask-AppBuilder==4.3.6
# # Flask-Babel==2.0.0
# # Flask-Caching==2.3.1
# # Flask-JWT-Extended==4.7.1
# # Flask-Limiter==3.10.1
# # Flask-Login==0.6.3
# # Flask-Session==0.5.0
# # Flask-SQLAlchemy==2.5.1
# # Flask-WTF==1.2.2
# flatten-dict==0.4.2
# flufl.lock==8.1.0
# frozenlist==1.5.0
# fsspec==2025.2.0
# funcy==2.0
# gcsfs==2025.2.0
# gitdb==4.0.12
# GitPython==3.1.44
# google-api-core==2.24.1
# google-auth==2.38.0
# google-auth-oauthlib==1.2.1
# google-cloud-core==2.4.2
# google-cloud-storage==3.1.0
# google-crc32c==1.6.0
# google-re2==1.1.20240702
# google-resumable-media==2.7.2
# googleapis-common-protos==1.67.0
# grandalf==0.8
# great-expectations==0.18.22
# greenlet==3.1.1
# grpcio==1.70.0
# gto==1.7.2
# gunicorn==23.0.0
# h11==0.14.0
# httpcore==1.0.7
# httpx==0.28.1
# huggingface-hub==0.28.1
# hydra-core==1.3.2
# idna==3.10
# importlib_metadata==8.5.0
# inflection==0.5.1
# iniconfig==2.0.0
# iterative-telemetry==0.0.10
# itsdangerous==2.2.0
# Jinja2==3.1.5
# jiter==0.8.2
# jmespath==1.0.1
# joblib==1.4.2
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# kombu==5.4.2
# langchain==0.3.18
# langchain-core==0.3.35
# langchain-huggingface==0.1.2
# langchain-pinecone==0.2.3
# langchain-tests==0.3.12
# langchain-text-splitters==0.3.6
# langsmith==0.3.8
# lazy-object-proxy==1.10.0
# limits==4.0.1
# linkify-it-py==2.0.3
# lockfile==0.12.2
# lz4==4.4.3
# Mako==1.3.9
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# marshmallow==3.26.1
# marshmallow-oneofschema==3.1.1
# marshmallow-sqlalchemy==0.28.2
# mdit-py-plugins==0.4.2
# mdurl==0.1.2
# methodtools==0.4.7
# mistune==3.1.2
# monotonic==1.6
# more-itertools==10.6.0
# mpmath==1.3.0
# multidict==6.1.0
# networkx==3.4.2
# #numpy==1.26.4
# oauthlib==3.2.2
# omegaconf==2.3.0
# openai==1.63.0
# opentelemetry-api==1.30.0
# opentelemetry-exporter-otlp==1.30.0
# opentelemetry-exporter-otlp-proto-common==1.30.0
# opentelemetry-exporter-otlp-proto-grpc==1.30.0
# opentelemetry-exporter-otlp-proto-http==1.30.0
# opentelemetry-proto==1.30.0
# opentelemetry-sdk==1.30.0
# opentelemetry-semantic-conventions==0.51b0
# ordered-set==4.1.0
# orjson==3.10.15
# packaging==24.2
# #pandas==2.1.4
# pathspec==0.12.1
# pendulum==2.1.2
# pillow==11.1.0
# pinecone==5.4.2
# pinecone-client==5.0.1
# pinecone-plugin-inference==1.1.0
# pinecone-plugin-interface==0.0.7
# platformdirs==4.3.6
# pluggy==1.5.0
# posthog==3.18.0
# prison==0.2.1
# prompt_toolkit==3.0.50
# propcache==0.2.1
# proto-plus==1.26.0
# protobuf==5.29.3
# protoc-gen-openapiv2==0.0.1
# psutil==7.0.0
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
# pycparser==2.22
# pydantic==2.10.6
# pydantic_core==2.27.2
# pydot==3.0.4
# pygit2==1.17.0
# Pygments==2.19.1
# pygtrie==2.5.0
# PyJWT==2.10.1
# pyparsing==3.2.1
# PyPDF2==3.0.1
# pytest==8.3.4
# pytest-asyncio==0.25.3
# pytest-socket==0.7.0
# python-daemon==3.1.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# python-nvd3==0.16.0
# python-slugify==8.0.4
# pytz==2025.1
# pywin32==308
# PyYAML==6.0.2
# referencing==0.36.2
# regex==2024.11.6
# requests==2.32.3
# requests-oauthlib==2.0.0
# requests-toolbelt==1.0.0
# rfc3339-validator==0.1.4
# rich==13.9.4
# rich-argparse==1.7.0
# rpds-py==0.23.1
# rsa==4.9
# ruamel.yaml==0.18.10
# ruamel.yaml.clib==0.2.12
# safetensors==0.5.2
# scikit-learn==1.6.1
# scipy==1.15.1
# scmrepo==3.3.10
# semver==3.0.4
# sentence-transformers==3.4.1
# setproctitle==1.3.5
# setuptools==75.8.0
# shellingham==1.5.4
# shortuuid==1.0.13
# shtab==1.7.1
# six==1.17.0
# smmap==5.0.2
# sniffio==1.3.1
# soupsieve==2.6
# SQLAlchemy==1.4.54
# SQLAlchemy-JSONField==1.0.2
# SQLAlchemy-Utils==0.41.2
# sqlparse==0.5.3
# sqltrie==0.11.2
# sympy==1.13.1
# syrupy==4.8.1
# tabulate==0.9.0
# tenacity==9.0.0
# termcolor==2.5.0
# text-unidecode==1.3
# threadpoolctl==3.5.0
# tiktoken==0.9.0
# time-machine==2.16.0
# tokenizers==0.21.0
# tomlkit==0.13.2
# toolz==1.0.0
# torch==2.6.0
# tqdm==4.67.1
# transformers==4.48.3
# typer==0.15.2
# typing_extensions==4.12.2
# tzdata==2025.1
# tzlocal==5.3
# uc-micro-py==1.0.3
# universal_pathlib==0.2.6
# urllib3==2.3.0
# vine==5.1.0
# voluptuous==0.15.2
# wcwidth==0.2.13
# Werkzeug==2.2.3
# wirerope==1.0.0
# wrapt==1.17.2
# WTForms==3.2.1
# yarl==1.18.3
# zc.lockfile==3.0.post1
# zipp==3.21.0
# zstandard==0.23.0


# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
openai==0.28.0
pinecone>=2.2.1  # Changed from pinecone-client to pinecone
sentence-transformers>=2.2.2

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# For Graph RAG
networkx>=2.8.0
pyvis>=0.3.1

# For evaluation
scikit-learn>=1.2.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Utilities
pydantic>=2.0.0
colorlog==4.7.2